feat(command): support for sub commands
Base command now subclasses the command registry making it possible to
have sub command. Command arguments (`foo bar baz --flag=x`) are treated
as lookup paths to anothe command to delegate to. If found it is
executed in the same way it was before.

There is no limit how deeply commands can be nested.

BREAKING CHANGE: Commands are no longer dot accessible from the regirstry
BREAKING CHANGE: The command registry is a Map, not an object