feat(seeli): add basic support for plugins
enable basic support for plugins. this allows for end users to develop
additional commands externally and load them dynamically.

a plugin can be an inline function or a string. if it is a string, it is
considered a require path and passed directly to require. The loaded
module should export a single function.

in both cases, the function is called with a null context and passed the
current instance of seeli

resolves: gh-39