fix(command): fix interactive prompts for repeating numbers
reconstruct arrays of numeric inputs into individual flag inputs for
nopt. this was working for strings because the values were technically
strings ('foo,bar,baz'). but for numeric flags, it was also being
constructed as a string ('1,2,3'), and nopt would fail validation.
This updates the interactive handler to reconstruct the value as
individual flags for nopt to validate