feat(command): add support for input filter for non-interactive
the input filter, previously supported for interactive mode via inquirer
allow transformation of the input before validation happens. This allows
for a great deal of functionality and flexibility for end users being
able to fix typos, enter short hands, etc. This is being extended into
the non-interactive commands as well