feat(flags)!: Add flags to help specify exclusivity
This commit adds various flags for mutual inclusion and exclusion. Doing
these kinds of things manually in `validate` usually breaks the
interactive mode. These options give the user clear ways to specify flag
dependencies.

BREAKING CHANGE: `interactive` values need to be reset from `''` to
`undefined` so that it behaves in the same way that missing argv values
do. Also, validation is trickier now with the new options, so a `major`
is the safe bet.